Fix ICC build error.
git-svn-id: https://edk2.svn.sourceforge.net/svnroot/edk2/trunk/edk2@11510 6f19259b-4bc3-4df7-8a09-765794883524

+++ MdeModulePkg/Universal/Network/IScsiDxe/IScsiConfig.c
@@ -1020,7 +1020,7 @@ IScsiConfigUpdateForm (
         //
         CopyMem (ConfigFormEntry->SessionConfigData.IsId, &MacAddress, 6);
         ConfigFormEntry->SessionConfigData.IsId[0] = 
-          ConfigFormEntry->SessionConfigData.IsId[0] & 0x3F;
+          (UINT8) (ConfigFormEntry->SessionConfigData.IsId[0] & 0x3F);
       }
       //
       // Get the CHAP authentication configuration data.
